Move Footer component into its own file to improve readability.
